refactor: use shared usePrefersReducedMotion hook
