--- components/hooks/usePrefersReducedMotion.js
@@ -1,16 +1,0 @@
-import { useEffect, useState } from 'react';
-
-// Hook to respect the user's reduced motion preference
-export default function usePrefersReducedMotion() {
-  const [prefersReduced, setPrefersReduced] = useState(false);
-
-  useEffect(() => {
-    const media = window.matchMedia('(prefers-reduced-motion: reduce)');
-    const handleChange = () => setPrefersReduced(media.matches);
-    handleChange();
-    media.addEventListener('change', handleChange);
-    return () => media.removeEventListener('change', handleChange);
-  }, []);
-
-  return prefersReduced;
-}
